Predictive task
The task was to classify incoming emails as spam or not. I identified the most relevant predictors and used the aggregate function to combine them into a single data set. Lift charts and a confusion matrix were then used for the validation data set to gather more information on the model. I continued the project by training a model that included all of the predictors, and then compared them. 

Various techniques were used to gather more information and finalize the model, including BIC, AIC, forward, backward, and step-wise selection.